refactoring: improving code quality
Moved if condition logic to switch because we are moving all of the code to switch.
The code only affects Aged Brie products.
All tests pass as it is, and we didn't modify any test.
Only two products are left to be moved to the switch block.
Afterwards, we will have a cleaner implementation of code.